Move platform-guessing from __init__.py to auto.py
See #70
