+++ tests/fake_sys_class/tacho-motor/motor0/state
@@ -1,1 +1,1 @@
-
+running
